[ADD/#1] Grayscale color 추가
