+++ app/src/main/java/org/sopt/and/designsystem/theme/Color.kt
@@ -1,0 +1,26 @@
+package org.sopt.and.designsystem.theme
+
+import androidx.compose.ui.graphics.Color
+
+val Purple80 = Color(0xFFD0BCFF)
+val PurpleGrey80 = Color(0xFFCCC2DC)
+val Pink80 = Color(0xFFEFB8C8)
+
+val Purple40 = Color(0xFF6650a4)
+val PurpleGrey40 = Color(0xFF625b71)
+val Pink40 = Color(0xFF7D5260)
+
+// Gray
+val White = Color(0xFFFFFFFF)
+val Gray50 = Color(0xFFFAFAFA)
+val Gray100 = Color(0xFFF5F5F5)
+val Gray200 = Color(0xFFE5E5E5)
+val Gray300 = Color(0xFFD6D6D6)
+val Gray350 = Color(0xFFC5C5C5)
+val Gray400 = Color(0xFFA6A6A6)
+val Gray500 = Color(0xFF737373)
+val Gray600 = Color(0xFF575757)
+val Gray700 = Color(0xFF424242)
+val Gray800 = Color(0xFF303030)
+val Gray850 = Color(0xFF1F1F1F)
+val Gray900 = Color(0xFF111111)
