+++ app/src/main/java/org/sopt/and/designsystem/theme/Theme.kt
@@ -1,6 +1,5 @@
-package org.sopt.and.ui.theme
+package org.sopt.and.designsystem.theme
 
-import android.app.Activity
 import android.os.Build
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material3.MaterialTheme
