+++ app/src/main/java/org/sopt/and/designsystem/theme/Type.kt
@@ -1,4 +1,4 @@
-package org.sopt.and.ui.theme
+package org.sopt.and.designsystem.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
