+++ app/src/main/java/org/sopt/and/feature/MainActivity.kt
@@ -1,4 +1,4 @@
-package org.sopt.and
+package org.sopt.and.feature
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -11,7 +11,7 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import org.sopt.and.ui.theme.ANDANDROIDTheme
+import org.sopt.and.designsystem.theme.ANDANDROIDTheme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
